pyproject.toml
[tool.poetry]
name = "signal-festival"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
flask = "^3.0.3"
flask-restx = "^1.3.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
aiogram = "^3.6.0"
aioflask = "^0.4.0"
fastapi = "^0.111.0"
openai = "^1.30.2"
tiktoken = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
